Wrap POD text more consistently
Much of the POD text was wrapped to within roughly 80 characters, whereas
some other parts of the text were very long single lines.  This wraps the
text to within 80 characters to be consistent with most of the text.
Wrapping the lines has further advantages in that it is easier to see in the
diffs where a given change took place, as well as allowing the text to fit
nicely in a standard terminal width.