Add link tags to App::TimeTracker instances in POD
The beginning of the tracker POD uses the link (L<>) formatting tags to link
to e.g. App::TimeTracker.  This commit extends this usage to the rest of the
document, so that the formatting is nice on metacpan and so that one can
follow references to App::TimeTracker later in the document.