Sort list of available commands for consistency
Subsequent calls of `tracker commands` would provide a different order
of the list of commands each time it was called.  This I found to be
confusing and it was difficult to find the command one was looking for.
Alphabetically sorting the list of available commands makes finding a
given command much easier.  I've also added a test for the `commands`
command to ensure the output is as expected.