Fix off-by-one rounding error in Task->rounded_minutes
The implementation always added an extra minute to the output, hence
when a task duration showed 00:15:00, the `rounded_minutes` value came
out at 16.  In the general case a duration of 00:x:00 would have a
`rounded_minutes` value of `x+1`.  The solution implemented here uses
a [hint from PerlMonks](https://www.perlmonks.org/bare/?node_id=257208),
which mentioned that to round to the nearest integer one must use
`sprintf "%.0f" $value` instead of `sprintf "%d" $value` because the
latter has the same truncation behaviour as `int()`.  This commit also
adds tests at the `Data::Task` level but also at the `Command::core`
level (which is where I noticed the discrepancy to start with).