Add an initial test for report command
It turned out that the `test` option in the test data `projects.json`
files was set to be the root directory (`/`).  This caused the output of
the `report` command not to be printed in the tests.  This is an
important part of this patch as without it the actual tests of the
`report` command won't work.

I've been able to check the `report` command output very generically in
the test added here, however it seems that even if one doesn't specify a
project, the output from the available projects get printed anyway
(hence the test for `report` output without a project is currently
commented out).  I'm not 100% sure at this stage as how to test this
more thoroughly, nevertheless, it *is* now possible to test the `report`
command :-)