Remove unused GitHub::Meta option
The reason why this option was unused is because it was spelled
incorrectly: it should be `homepage`, not `hompage` (not the missing
'e').  However, beyond that, the `homepage` metadata value was being
overridden by the `MetaResources` plugin anyway, hence the setting
`homepage=false` would have no (visible) effect; in other words it can
be removed without changing the desired values of the dist metadata.