Test current start command help output without project
Before extending the help output given when the `start` command is used
without having a current project, it's necessary to ensure that the
current behaviour is tested, hence this commit.

The background behind this change is RT#88687, where Gabor Szabo noticed
that if one didn't have a project and tried to start work (via
`tracker start`), then the help output wasn't very helpful and it should
probably mention that one needs to initialise the project first with
`tracker init`.