Fix Dist::Zilla build problems
In order to run `dzil authordeps --missing | cpanm` to completion, it
was necessary to add `OurPkgVersion` to the list of required plugins.
Afterwards, `dzil test` and `dzil listdeps` failed to run due to missing
`license` and `copyright_holder` fields.  These fields have been added
and it is possible to build and test the dist again.